Remove weird things from the Debug configuration.
Since Debug is apparently the default in MonoDevelop and is required for
running the tests, it should be a configuration suitable for development.
That means traces (which cause the tests to break) and IronAHK.Program.Start
(which causes my brain to break) are no good. If these are really needed, they
should be in another configuration that defines TRACE or WTF respectively.